fix: BoardPage 구조 변경
